rename the new map component, delete the old map component and amend the contact block
