+++ docs/components/contacts-block.md
@@ -106,12 +106,13 @@ import InitAll from "../../src/InitAll"
 </ul>
 </div>
 </div>
-<div class="lbh-contact__map-container">
-<div class="lbh-contact__map" id="map" data-module="lbh-map" data-access-token="pk.eyJ1IjoibGJoZWxld2lzIiwiYSI6ImNqeXJkN25uNjA5M3Uzb251bWVyejJ3YW8ifQ.uzO8I54w64U6QkNknW32FA" data-marker-lat="51.545386" data-marker-lng="-0.057069"></div>
-<div class="lbh-contact__directions">
+<div class="lbh-map__map-container">
+<div class="lbh-map__map" data-module="lbh-map" data-marker-lat="51.545386" data-marker-lng="-0.057069" uprn="100023014187"></div>
+<div class="lbh-map__directions">
 <a href="https://goo.gl/maps/YHtCx2nqP2o57BZi6" class="lbh-link" rel="external" title="View directions on Google Maps">Get directions</a>
 </div>
 </div>
+
 </section>
 
 </InitAll>
@@ -127,6 +128,8 @@ import L from "leaflet"
 ```scss
 @import "../node_modules/leaflet/dist/leaflet.css";
 ```
+
+Please email us at gis@hackney.gov.uk to get the OS basemap token and the addresses proxy url for your project.
 ### HTML
 
 ```html
@@ -356,16 +359,16 @@ import L from "leaflet"
       </ul>
     </div>
   </div>
-  <div class="lbh-contact__map-container">
+  <div class="lbh-map__map-container">
     <div
-      class="lbh-contact__map"
+      class="lbh-map__map"
       id="map"
       data-module="lbh-map"
-      data-access-token="pk.eyJ1IjoibGJoZWxld2lzIiwiYSI6ImNqeXJkN25uNjA5M3Uzb251bWVyejJ3YW8ifQ.uzO8I54w64U6QkNknW32FA"
       data-marker-lat="51.545386"
       data-marker-lng="-0.057069"
+      uprn="100023014187"
     ></div>
-    <div class="lbh-contact__directions">
+    <div class="lbh-map__directions">
       <a
         href="https://goo.gl/maps/YHtCx2nqP2o57BZi6"
         class="lbh-link"
@@ -374,6 +377,7 @@ import L from "leaflet"
         >Get directions</a
       >
     </div>
+  </div>
   </div>
 </section>
 ```
