+++ docs/components/map.md
@@ -6,9 +6,9 @@ import InitAll from "../../src/InitAll"
 
 <InitAll>
 
-<section class="lbh-map2">
-<div class="lbh-map__map-container">
-<div class="lbh-map__map" data-module="lbh-map2" data-marker-lat="51.545386" data-marker-lng="-0.057069" uprn="100023014187"></div>
+<section class="lbh-map">
+<div class="lbh-map__map-container_full_width">
+<div class="lbh-map__map" data-module="lbh-map" data-marker-lat="51.545386" data-marker-lng="-0.057069" uprn="100023014187"></div>
 <div class="lbh-map__directions">
 <a href="https://goo.gl/maps/YHtCx2nqP2o57BZi6" class="lbh-link" rel="external" title="View directions on Google Maps">Get directions</a>
 </div>
@@ -32,12 +32,12 @@ Please email us at gis@hackney.gov.uk to get the OS basemap token and the addres
 ### HTML
 
 ```html
-<section class="lbh-map2"> 
-  <div class="lbh-map__map-container">
+<section class="lbh-map"> 
+  <div class="lbh-map__map-container_full_width">
     <div
       class="lbh-map__map"
-      id="map2"
-      data-module="lbh-map2"
+      id="map"
+      data-module="lbh-map"
       data-marker-lat="51.545386"
       data-marker-lng="-0.057069"
       uprn="100023014187"
