+++ lbh/components/_all.scss
@@ -32,7 +32,7 @@
 @import "lbh-tag/tag";
 @import "lbh-textarea/textarea";
 @import "lbh-warning-text/warning-text";
-@import "lbh-map2/map";
+@import "lbh-map/map";
 
 // staff-facing components
 @import "lbh-collapsible/collapsible";
